requirements.txt
# ============================================================
#  Modern Agentic RAG — requirements.txt
#  Install: pip install -r requirements.txt
#  GPU (CUDA 12.x): add --extra-index-url https://download.pytorch.org/whl/cu121
# ============================================================

# ── Core ─────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
tqdm>=4.66.0

# ── PDF Parsing ───────────────────────────────────────────────────────────
# High-quality parser (recommended)
unstructured[pdf]>=0.14.0
# Fast fallback
PyMuPDF>=1.24.0

# ── NLP ───────────────────────────────────────────────────────────────────
spacy>=3.7.0
# python -m spacy download en_core_web_sm  (run once after install)

# ── Embeddings ────────────────────────────────────────────────────────────
sentence-transformers>=3.0.0
torch>=2.2.0                  # install GPU version separately if needed

# ── Vector Store ──────────────────────────────────────────────────────────
qdrant-client>=1.9.0

# ── Hybrid Search ─────────────────────────────────────────────────────────
rank-bm25>=0.2.2

# ── Knowledge Graph ───────────────────────────────────────────────────────
networkx>=3.3

# ── LangGraph (agentic loop) ──────────────────────────────────────────────
langgraph>=0.2.0
langchain-core>=0.3.0

# ── LLM Providers (install only what you need) ───────────────────────────
anthropic>=0.34.0             # Claude
openai>=1.40.0                # GPT-4o
google-generativeai>=0.8.0    # Gemini

# LangChain adapters for RAGAs judge LLM
langchain-anthropic>=0.2.0
langchain-openai>=0.2.0
langchain-google-genai>=2.0.0

# ── Local HuggingFace LLMs (optional) ─────────────────────────────────────
transformers>=4.44.0
accelerate>=0.33.0
bitsandbytes>=0.43.0          # 4-bit quantization (Linux/GPU only)

# ── Evaluation ────────────────────────────────────────────────────────────
ragas>=0.2.0
datasets>=2.20.0

# ── Dev / Notebook (optional) ────────────────────────────────────────────
jupyter>=1.0.0
ipykernel>=6.29.0
